Some things

+++ src/Server.cpp
@@ -3,7 +3,6 @@
 #include <chrono>
 #include <csignal>
 #include <cstring>
-#include <iomanip>
 #include <iostream>
 #include <memory>
 #include <ostream>
@@ -261,7 +260,7 @@ bool Server::_init() noexcept {
         return false;
     }
 
-    std::cout << "Server is running on: " << _port << ". Press Ctrl+C to stop."
+    std::cout << "Server is running on: " << static_cast<unsigned int>(_port) << ". Press Ctrl+C to stop."
               << '\n';
     return true;
 }
